+++ server/src/main/java/com/epam/aidial/core/server/validation/BeanDeserializerModifierWithValidation.java
@@ -1,11 +1,17 @@
 package com.epam.aidial.core.server.validation;
 
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.BeanDescription;
 import com.fasterxml.jackson.databind.DeserializationConfig;
+import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.deser.BeanDeserializer;
 import com.fasterxml.jackson.databind.deser.BeanDeserializerBase;
 import com.fasterxml.jackson.databind.deser.BeanDeserializerModifier;
+
+import java.io.IOException;
+
+import static com.epam.aidial.core.server.validation.ValidationUtil.validate;
 
 public class BeanDeserializerModifierWithValidation extends BeanDeserializerModifier {
     @Override
@@ -15,4 +21,20 @@ public class BeanDeserializerModifierWithValidation extends BeanDeserializerModi
         }
         return deserializer;
     }
+
+    private static class BeanDeserializerWithValidation extends BeanDeserializer {
+
+        protected BeanDeserializerWithValidation(BeanDeserializerBase src) {
+            super(src);
+        }
+
+        @Override
+        public Object deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
+            Object instance = super.deserialize(p, ctxt);
+            validate(instance);
+            return instance;
+        }
+
+    }
+
 }
