--- server/src/main/java/com/epam/aidial/core/server/validation/BeanDeserializerWithValidation.java
@@ -1,26 +1,0 @@
-package com.epam.aidial.core.server.validation;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.deser.BeanDeserializer;
-import com.fasterxml.jackson.databind.deser.BeanDeserializerBase;
-
-import java.io.IOException;
-
-import static com.epam.aidial.core.server.validation.ValidationUtil.validate;
-
-
-public class BeanDeserializerWithValidation extends BeanDeserializer {
-
-    protected BeanDeserializerWithValidation(BeanDeserializerBase src) {
-        super(src);
-    }
-
-    @Override
-    public Object deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
-        Object instance = super.deserialize(p, ctxt);
-        validate(instance);
-        return instance;
-    }
-
-}
